finalização backend

+++ backend/src/repositories/transactionRepository.js
@@ -1,27 +1,27 @@
 import pool from '../config/database.js'
 
-export async function findAllTransactions() {
-  const result = await pool.query(`
-    SELECT
-      transactions.id,
-      transactions.user_id,
-      transactions.category_id,
-      transactions.description,
-      transactions.amount,
-      transactions.type,
-      transactions.transaction_date,
-      transactions.created_at,
-      transactions.updated_at,
-
-      categories.name AS category_name
-
-    FROM transactions
-
-    LEFT JOIN categories
-      ON transactions.category_id = categories.id
-
-    ORDER BY transactions.transaction_date DESC
-  `)
+export async function findAllTransactions(userId) {
+  const result = await pool.query(
+    `
+      SELECT
+        transactions.id,
+        transactions.user_id,
+        transactions.category_id,
+        transactions.description,
+        transactions.amount,
+        transactions.type,
+        transactions.transaction_date,
+        transactions.created_at,
+        transactions.updated_at,
+        categories.name AS category_name
+      FROM transactions
+      LEFT JOIN categories
+        ON transactions.category_id = categories.id
+      WHERE transactions.user_id = $1
+      ORDER BY transactions.transaction_date DESC
+    `,
+    [userId]
+  )
 
   return result.rows
 }
@@ -71,7 +71,11 @@ export async function createTransaction(data) {
   return result.rows[0]
 }
 
-export async function updateTransaction(id, data) {
+export async function updateTransaction(
+  id,
+  userId,
+  data
+) {
   const {
     category_id,
     description,
@@ -91,6 +95,7 @@ export async function updateTransaction(id, data) {
         transaction_date = $5,
         updated_at = NOW()
       WHERE id = $6
+        AND user_id = $7
       RETURNING
         id,
         user_id,
@@ -109,20 +114,25 @@ export async function updateTransaction(id, data) {
       type,
       transaction_date,
       id,
+      userId,
     ]
   )
 
   return result.rows[0]
 }
 
-export async function deleteTransaction(id) {
+export async function deleteTransaction(
+  id,
+  userId
+) {
   const result = await pool.query(
     `
       DELETE FROM transactions
       WHERE id = $1
+        AND user_id = $2
       RETURNING id
     `,
-    [id]
+    [id, userId]
   )
 
   return result.rows[0]
